requirements.txt
streamlit
python-dotenv
requests
faiss-cpu
sentence-transformers
pymupdf
numpy
tqdm
